Rmove requests helper, move parser one level up
