+++ meeseeks/parser/parser_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	parser "github.com/gomeeseeks/meeseeks-box/meeseeks/request/parser"
+	parser "github.com/gomeeseeks/meeseeks-box/meeseeks/parser"
 )
 
 func Test_ParsingCommandsCorrectly(t *testing.T) {
